Split monster pghkp/storage.go into smaller files; camelCasify fetchKeyDocs
