+++ src/hockeypuck/hkp/storage/storage.go
@@ -35,6 +35,8 @@ func IsNotFound(err error) bool {
 	return errors.Is(err, ErrKeyNotFound)
 }
 
+// Record is a PrimaryKey annotated with selected fields returned by the DB layer.
+// It is not a faithful representation of the underlying DB schema.
 type Record struct {
 	*openpgp.PrimaryKey
