+++ src/hockeypuck/pghkp/storage.go
@@ -1,6 +1,6 @@
 /*
    Hockeypuck - OpenPGP key server
-   Copyright (C) 2012-2014  Casey Marshall
+   Copyright (C) 2012-2025 Casey Marshall and the Hockeypuck Contributors
 
    This program is free software: you can redistribute it and/or modify
    it under the terms of the GNU Affero General Public License as published by
@@ -20,11 +20,7 @@ package pghkp
 import (
 	"bytes"
 	"database/sql"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"iter"
-	"maps"
 	"strings"
 	"sync"
 	"time"
@@ -37,8 +33,6 @@ import (
 	"github.com/pkg/errors"
 	"golang.org/x/exp/slices"
 
-	"hockeypuck/hkp/jsonhkp"
-	pksstorage "hockeypuck/hkp/pks/storage"
 	hkpstorage "hockeypuck/hkp/storage"
 	"hockeypuck/openpgp"
 
@@ -131,221 +125,6 @@ var drConstraintsSQL = []string{
 	`DROP INDEX subkeys_rfp;`,
 }
 
-var crTempTablesSQL = []string{
-	`CREATE TEMPORARY TABLE IF NOT EXISTS keys_copyin
-(
-rfingerprint TEXT,
-doc jsonb,
-ctime TIMESTAMPTZ,
-mtime TIMESTAMPTZ,
-idxtime TIMESTAMPTZ,
-md5 TEXT,
-keywords tsvector
-)
-`,
-	`CREATE TEMPORARY TABLE IF NOT EXISTS subkeys_copyin
-(
-rfingerprint TEXT,
-rsubfp TEXT
-)
-`,
-	`CREATE TEMPORARY TABLE IF NOT EXISTS keys_checked
-(
-rfingerprint TEXT NOT NULL PRIMARY KEY,
-doc jsonb NOT NULL,
-ctime TIMESTAMPTZ NOT NULL,
-mtime TIMESTAMPTZ NOT NULL,
-idxtime TIMESTAMPTZ NOT NULL,
-md5 TEXT NOT NULL UNIQUE,
-keywords tsvector
-)
-`,
-	`CREATE TEMPORARY TABLE IF NOT EXISTS subkeys_checked
-(
-rfingerprint TEXT NOT NULL,
-rsubfp TEXT NOT NULL PRIMARY KEY
-)
-`,
-}
-
-var drTempTablesSQL = []string{
-	`DROP TABLE IF EXISTS subkeys_copyin CASCADE
-`,
-	`DROP TABLE IF EXISTS keys_copyin CASCADE
-`,
-	`DROP TABLE IF EXISTS subkeys_checked CASCADE
-`,
-	`DROP TABLE IF EXISTS keys_checked CASCADE
-`,
-}
-
-// bulkTxFilterUniqueKeys is a key-filtering query, between temporary tables, used for bulk insertion.
-// Among all the keys in a call to Insert(..) (usually the keys in a processed key-dump file), this
-// filter gets the unique keys, i.e., those with unique rfingerprint *and* unique md5, but *neither*
-// with rfingerprint *nor* with md5 that currently exist in the DB.
-const bulkTxFilterUniqueKeys string = `INSERT INTO keys_checked (rfingerprint, doc, ctime, mtime, idxtime, md5, keywords) 
-SELECT rfingerprint, doc, ctime, mtime, idxtime, md5, keywords FROM keys_copyin kcpinA WHERE 
-rfingerprint IS NOT NULL AND doc IS NOT NULL AND ctime IS NOT NULL AND mtime IS NOT NULL AND idxtime IS NOT NULL AND md5 IS NOT NULL AND 
-(SELECT COUNT (*) FROM keys_copyin kcpinB WHERE kcpinB.rfingerprint = kcpinA.rfingerprint OR 
-                                                kcpinB.md5          = kcpinA.md5) = 1 AND 
-NOT EXISTS (SELECT 1 FROM keys WHERE keys.rfingerprint = kcpinA.rfingerprint OR keys.md5 = kcpinA.md5)
-`
-
-// bulkTxPrepKeyStats is a key-processing query on bulk insertion temporary tables that facilitates
-// calculation of statistics on keys and subsequent additional filtering. Out of all the keys in a
-// call to Insert(..) (usually the keys in a processed key-dump file), this query keeps only duplicates
-// by dropping keys previously set aside by bulkTxFilterUniqueKeys query and removing any tuples
-// with NULLs.
-const bulkTxPrepKeyStats string = `DELETE FROM keys_copyin WHERE 
-rfingerprint IS NULL OR doc IS NULL OR ctime IS NULL OR mtime IS NULL OR idxtime IS NULL OR md5 IS NULL OR 
-EXISTS (SELECT 1 FROM keys_checked WHERE keys_checked.rfingerprint = keys_copyin.rfingerprint)
-`
-
-// bulkTxFilterDupKeys is the final key-filtering query, between temporary tables, used for bulk
-// insertion. Among all the keys in a call to Insert(..) (usually the keys in a processed key-dump
-// file), this query sets aside for final DB insertion _a single copy_ of those keys that are
-// duplicates in the arguments of Insert(..), but do not yet exist in the DB.
-const bulkTxFilterDupKeys string =
-// *** ctid field is PostgreSQL-specific; Oracle has ROWID equivalent field ***
-// ===> If there are different md5 for same rfp, this query allows them into keys_checked: <===
-// ===>  ***  an intentional error of non-unique rfp, to revert to normal insertion!  ***  <===
-`INSERT INTO keys_checked (rfingerprint, doc, ctime, mtime, idxtime, md5, keywords) 
-SELECT rfingerprint, doc, ctime, mtime, idxtime, md5, keywords FROM keys_copyin WHERE 
-( ctid IN 
-     (SELECT ctid FROM 
-        (SELECT ctid, ROW_NUMBER() OVER (PARTITION BY rfingerprint ORDER BY ctid) rfpEnum FROM keys_copyin) AS dupRfpTAB 
-        WHERE rfpEnum = 1) OR 
-  ctid IN 
-     (SELECT ctid FROM 
-        (SELECT ctid, ROW_NUMBER() OVER (PARTITION BY md5 ORDER BY ctid) md5Enum FROM keys_copyin) AS dupMd5TAB 
-        WHERE md5Enum = 1) ) AND 
-NOT EXISTS (SELECT 1 FROM keys WHERE keys.rfingerprint = keys_copyin.rfingerprint OR
-                                     keys.md5          = keys_copyin.md5)
-`
-
-// bulkTxFilterUniqueSubkeys is a subkey-filtering query, between temporary tables, used for bulk
-// insertion. Among all the subkeys of keys in a call to Insert(..) (usually the keys in a processed
-// key-dump file), this filter gets the unique subkeys, i.e., those with no NULL fields that are not
-// duplicates (unique among subkeys of keys in this call to Insert(..) that do not currently exist in the DB).
-const bulkTxFilterUniqueSubkeys string =
-// Enforce foreign key constraint by checking both keys_checked and keys_copyin (instead of keys)
-// Avoid checking "EXISTS (SELECT 1 FROM keys WHERE keys.rfingerprint = skcpinA.rfingerprint)"
-// by checking in keys_copyin (despite no indexing): only duplicates (in-file or _in DB_) are
-// still in keys_copyin
-`INSERT INTO subkeys_checked (rfingerprint, rsubfp) 
-SELECT rfingerprint, rsubfp FROM subkeys_copyin skcpinA WHERE 
-skcpinA.rfingerprint IS NOT NULL AND skcpinA.rsubfp IS NOT NULL AND 
-(SELECT COUNT(*) FROM subkeys_copyin skcpinB WHERE skcpinB.rsubfp = skcpinA.rsubfp) = 1 AND 
-NOT EXISTS (SELECT 1 FROM subkeys WHERE subkeys.rsubfp = skcpinA.rsubfp) AND 
-( EXISTS (SELECT 1 FROM keys_checked WHERE keys_checked.rfingerprint = skcpinA.rfingerprint) OR 
-  EXISTS (SELECT 1 FROM keys_copyin  WHERE keys_copyin.rfingerprint  = skcpinA.rfingerprint) )
-`
-
-// bulkTxPrepSubkeyStats is a subkey-processing query on bulk insertion temporary tables that
-// facilitates calculation of statistics on subkeys and subsequent additional filtering. Out of
-// all the subkeys of keys in a call to Insert(..) (usually the keys in a processed key-dump file),
-// this query keeps only duplicates by dropping subkeys previously set aside by bulkTxFilterUniqueSubkeys
-// query and removing any tuples with NULLs.
-const bulkTxPrepSubkeyStats string = `DELETE FROM subkeys_copyin WHERE 
-rfingerprint IS NULL OR rsubfp IS NULL OR 
-EXISTS (SELECT 1 FROM subkeys_checked WHERE subkeys_checked.rsubfp = subkeys_copyin.rsubfp)
-`
-
-// bulkTxFilterDupSubkeys is the final subkey-filtering query, between temporary tables, used for
-// bulk insertion. Among all the subkeys of keys in a call to Insert(..) (usually the keys in a processed
-// key-dump file), this query sets aside for final DB insertion _a single copy_ of those subkeys that are
-// duplicates in the arguments of Insert(..), but do not yet exist in the DB.
-const bulkTxFilterDupSubkeys string =
-// Enforce foreign key constraint by checking both keys_checked and keys_copyin (instead of keys)
-// *** ctid field is PostgreSQL-specific; Oracle has ROWID equivalent field ***
-// Avoid checking "EXISTS (SELECT 1 FROM keys WHERE keys.rfingerprint = subkeys_copyin.rfingerprint)"
-// by checking in keys_copyin (despite no indexing): only dups (in-file or _in DB_) still in keys_copyin
-`INSERT INTO subkeys_checked (rfingerprint, rsubfp) 
-SELECT rfingerprint, rsubfp FROM subkeys_copyin WHERE 
-ctid IN 
-   (SELECT ctid FROM 
-      (SELECT ctid, ROW_NUMBER() OVER (PARTITION BY rsubfp ORDER BY ctid) rsubfpEnum FROM subkeys_copyin) AS dupRsubfpTAB 
-      WHERE rsubfpEnum = 1) AND 
-NOT EXISTS (SELECT 1 FROM subkeys WHERE subkeys.rsubfp = subkeys_copyin.rsubfp) AND 
-( EXISTS (SELECT 1 FROM keys_checked WHERE keys_checked.rfingerprint = subkeys_copyin.rfingerprint) OR 
-  EXISTS (SELECT 1 FROM keys_copyin  WHERE keys_copyin.rfingerprint  = subkeys_copyin.rfingerprint) )
-`
-
-// bulkTxInsertKeys is the query for final bulk key insertion, from a temporary table to the DB.
-const bulkTxInsertKeys string = `INSERT INTO keys (rfingerprint, doc, ctime, mtime, idxtime, md5, keywords) 
-SELECT rfingerprint, doc, ctime, mtime, idxtime, md5, keywords FROM keys_checked
-`
-
-// bulkTxInsertSubkeys is the query for final bulk subkey insertion, from a temporary table to the DB.
-const bulkTxInsertSubkeys string = `INSERT INTO subkeys (rfingerprint, rsubfp) 
-SELECT rfingerprint, rsubfp FROM subkeys_checked
-`
-
-// bulkTxReindexKeys is the query for updating the SQL schema only, from a temporary table to the DB.
-// We match on the md5 field only, to prevent race conditions (this is safe since md5 is UNIQUE).
-const bulkTxReindexKeys string = `UPDATE keys
-SET idxtime = keys_copyin.idxtime, keywords = keys_copyin.keywords FROM keys_copyin
-WHERE keys.md5 = keys_copyin.md5
-`
-
-// Stats collection queries
-
-const bulkInsNumNullKeys string = `SELECT COUNT (*) FROM keys_copyin WHERE 
-rfingerprint IS NULL OR doc IS NULL OR ctime IS NULL OR mtime IS NULL OR idxtime IS NULL OR md5 IS NULL
-`
-const bulkInsNumNullSubkeys string = `SELECT COUNT (*) FROM subkeys_copyin WHERE 
-rfingerprint IS NULL OR rsubfp IS NULL
-`
-const bulkInsNumMinDups string = `SELECT COUNT (*) FROM keys_copyin WHERE 
-( ( NOT EXISTS (SELECT 1 FROM keys_checked WHERE keys_checked.rfingerprint = keys_copyin.rfingerprint OR
-                                                 keys_checked.md5          = keys_copyin.md5) AND 
-    ctid IN 
-       (SELECT ctid FROM 
-          (SELECT ctid, ROW_NUMBER() OVER (PARTITION BY rfingerprint ORDER BY ctid) rfpEnum FROM keys_copyin) AS dupRfpTAB 
-          WHERE rfpEnum = 1) ) OR 
-  ctid IN 
-     (SELECT ctid FROM 
-        (SELECT ctid, ROW_NUMBER() OVER (PARTITION BY rfingerprint) rfpEnum FROM keys_copyin) AS dupRfpTAB 
-        WHERE rfpEnum > 1) ) AND 
-NOT EXISTS (SELECT 1 FROM subkeys_checked WHERE subkeys_checked.rfingerprint = keys_copyin.rfingerprint)
-`
-const bulkInsNumPossibleDups string = `SELECT COUNT (*) FROM keys_copyin WHERE 
-ctid IN 
-   (SELECT ctid FROM 
-      (SELECT ctid, ROW_NUMBER() OVER (PARTITION BY rfingerprint) rfpEnum FROM keys_copyin) AS dupRfpTAB 
-      WHERE rfpEnum > 1) AND 
-EXISTS (SELECT 1 FROM subkeys_checked WHERE subkeys_checked.rfingerprint = keys_copyin.rfingerprint)
-`
-const bulkInsertedKeysNum string = `SELECT COUNT (*) FROM keys_checked
-`
-const bulkInsertedSubkeysNum string = `SELECT COUNT (*) FROM subkeys_checked
-`
-const bulkCopiedKeysNum string = `SELECT COUNT (*) FROM keys_copyin
-`
-
-const bulkInsQueryKeyChange string = `SELECT md5 FROM keys_checked
-`
-
-const keys_copyin_temp_table_name string = "keys_copyin"
-const subkeys_copyin_temp_table_name string = "subkeys_copyin"
-
-// keysInBunch is the maximum number of keys sent in a bunch during bulk insertion.
-// Since keys (and subkeys) are sent to the DB in prepared statements with parameters and
-// each key requires keysNumColumns parameters, keysInBunch < 65536/keysNumColumns.
-// 64k (2-byte parameter count) is the current protocol limit for client communication,
-// of prepared statements in PostreSQL v13 (see Bind message in
-// https://www.postgresql.org/docs/current/protocol-message-formats.html).
-const keysInBunch int = 64000 / keysNumColumns
-
-// subkeysInBunch is the maximum number of subkeys sent in a bunch (for at most
-// keysInBunch keys sent in a bunch) during bulk insertion. If each subkey requires 2
-// parameters, ~32k subkeys can fit in a bunch (see keysInBunch).
-const subkeysInBunch int = 64000 / subkeysNumColumns
-
-// minKeys2UseBulk is the minimum number of keys in a call to Insert(..) that
-// will trigger a bulk insertion. Otherwise, Insert(..) preceeds one key at a time.
-const minKeys2UseBulk int = 3500
-
 // Dial returns PostgreSQL storage connected to the given database URL.
 func Dial(url string, options []openpgp.KeyReaderOption) (hkpstorage.Storage, error) {
 	db, err := sql.Open("postgres", url)
@@ -399,6 +178,7 @@ func (st *storage) createIndexes() error {
 	return nil
 }
 
+// keyDoc is a nearly-raw copy of a row in the PostgreSQL `keys` table.
 type keyDoc struct {
 	RFingerprint string
 	CTime        time.Time
@@ -407,366 +187,6 @@ type keyDoc struct {
 	MD5          string
 	Doc          string
 	Keywords     string
-}
-
-func (st *storage) MatchMD5(md5s []string) ([]string, error) {
-	var md5In []string
-	var md5Values []string
-	for _, md5 := range md5s {
-		// Must validate to prevent SQL injection since we're appending SQL strings here.
-		_, err := hex.DecodeString(md5)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid MD5 %q", md5)
-		}
-		md5In = append(md5In, "'"+strings.ToLower(md5)+"'")
-		md5Values = append(md5Values, "('"+strings.ToLower(md5)+"')")
-	}
-
-	sqlStr := fmt.Sprintf("SELECT rfingerprint FROM keys WHERE md5 IN (%s)", strings.Join(md5In, ","))
-	rows, err := st.Query(sqlStr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	var result []string
-	defer rows.Close()
-	for rows.Next() {
-		var rfp string
-		err := rows.Scan(&rfp)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		result = append(result, rfp)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	// If we receive a hashquery for nonexistent digest(s), assume the ptree is stale and force an update.
-	// https://github.com/hockeypuck/hockeypuck/issues/170#issuecomment-1384003238 (note 1)
-	sqlStr = fmt.Sprintf("SELECT md5 FROM (values %s) as hashquery(md5) WHERE NOT EXISTS (SELECT FROM keys WHERE md5 = hashquery.md5)", strings.Join(md5Values, ","))
-	rows, err = st.Query(sqlStr)
-	if err == nil {
-		for rows.Next() {
-			var md5 string
-			err := rows.Scan(&md5)
-			if err == nil {
-				st.Notify(hkpstorage.KeyRemovedJitter{ID: "??", Digest: md5})
-			}
-		}
-	}
-
-	return result, nil
-}
-
-// Resolve implements storage.Storage.
-//
-// Only v4 key IDs are resolved by this backend. v3 short and long key IDs
-// currently won't match.
-func (st *storage) Resolve(keyids []string) (_ []string, retErr error) {
-	var result []string
-	sqlStr := "SELECT rfingerprint FROM keys WHERE rfingerprint LIKE $1 || '%'"
-	stmt, err := st.Prepare(sqlStr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	defer stmt.Close()
-
-	var subKeyIDs []string
-	for _, keyid := range keyids {
-		keyid = strings.ToLower(keyid)
-		var rfp string
-		row := stmt.QueryRow(keyid)
-		err = row.Scan(&rfp)
-		if err == sql.ErrNoRows {
-			subKeyIDs = append(subKeyIDs, keyid)
-		} else if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		result = append(result, rfp)
-	}
-
-	if len(subKeyIDs) > 0 {
-		subKeyResult, err := st.resolveSubKeys(subKeyIDs)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		result = append(result, subKeyResult...)
-	}
-
-	return result, nil
-}
-
-func (st *storage) resolveSubKeys(keyids []string) ([]string, error) {
-	var result []string
-	sqlStr := "SELECT rfingerprint FROM subkeys WHERE rsubfp LIKE $1 || '%'"
-	stmt, err := st.Prepare(sqlStr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	defer stmt.Close()
-
-	for _, keyid := range keyids {
-		keyid = strings.ToLower(keyid)
-		var rfp string
-		row := stmt.QueryRow(keyid)
-		err = row.Scan(&rfp)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		result = append(result, rfp)
-	}
-
-	return result, nil
-}
-
-func (st *storage) MatchKeyword(search []string) ([]string, error) {
-	var result []string
-	stmt, err := st.Prepare("SELECT rfingerprint FROM keys WHERE keywords @@ $1::TSQUERY LIMIT $2")
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	defer stmt.Close()
-
-	for _, term := range search {
-		err = func() error {
-			query, err := keywordsTSQuery(term)
-			if err != nil {
-				return errors.WithStack(err)
-			}
-			rows, err := stmt.Query(query, 100)
-			if err != nil {
-				return errors.WithStack(err)
-			}
-			defer rows.Close()
-			for rows.Next() {
-				var rfp string
-				err = rows.Scan(&rfp)
-				if err != nil && err != sql.ErrNoRows {
-					return errors.WithStack(err)
-				}
-				result = append(result, rfp)
-			}
-			err = rows.Err()
-			if err != nil {
-				return errors.WithStack(err)
-			}
-			return nil
-		}()
-		if err != nil {
-			return nil, err
-		}
-	}
-	return result, nil
-}
-
-func (st *storage) ModifiedSince(t time.Time) ([]string, error) {
-	var result []string
-	rows, err := st.Query("SELECT rfingerprint FROM keys WHERE mtime > $1 ORDER BY mtime ASC LIMIT 100", t.UTC())
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	defer rows.Close()
-	for rows.Next() {
-		var rfp string
-		err = rows.Scan(&rfp)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		result = append(result, rfp)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	return result, nil
-}
-
-var deletedKey = errors.Errorf("key deleted")
-
-func (st *storage) preen(key *openpgp.PrimaryKey, pk jsonhkp.PrimaryKey, sqlMD5 string) error {
-	if key == nil {
-		log.Warnf("unparseable key material in database (fp=%s); deleting", pk.Fingerprint)
-		_, err := st.Delete(pk.Fingerprint)
-		if err != nil {
-			log.Errorf("could not delete fp=%s", pk.Fingerprint)
-			return err
-		}
-		return deletedKey
-	}
-	if len(key.SubKeys) == 0 && len(key.UserIDs) == 0 && len(key.Signatures) == 0 {
-		log.Warnf("lone primary key packet in database (fp=%s); deleting", pk.Fingerprint)
-		_, err := st.Delete(pk.Fingerprint)
-		if err != nil {
-			log.Errorf("could not delete fp=%s", pk.Fingerprint)
-			return err
-		}
-		return deletedKey
-	}
-	if key.MD5 != sqlMD5 {
-		log.Warnf("MD5 changed while parsing (old=%s new=%s fp=%s); cleaning", sqlMD5, key.MD5, pk.Fingerprint)
-		// Beware this may cause double-updates in some circumstances
-		err := st.Update(key, key.KeyID(), sqlMD5)
-		if err != nil {
-			log.Errorf("could not clean fp=%s", pk.Fingerprint)
-			return err
-		}
-	}
-	return nil
-}
-
-func (st *storage) FetchKeys(rfps []string, options ...string) ([]*openpgp.PrimaryKey, error) {
-	autoPreen := slices.Contains(options, hkpstorage.AutoPreen)
-	if len(rfps) == 0 {
-		return nil, nil
-	}
-
-	var rfpIn []string
-	for _, rfp := range rfps {
-		_, err := hex.DecodeString(rfp)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid rfingerprint %q", rfp)
-		}
-		rfpIn = append(rfpIn, "'"+strings.ToLower(rfp)+"'")
-	}
-	sqlStr := fmt.Sprintf("SELECT doc, md5 FROM keys WHERE rfingerprint IN (%s)", strings.Join(rfpIn, ","))
-	rows, err := st.Query(sqlStr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	var result []*openpgp.PrimaryKey
-	defer rows.Close()
-	for rows.Next() {
-		var bufStr, sqlMD5 string
-		err = rows.Scan(&bufStr, &sqlMD5)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		var pk jsonhkp.PrimaryKey
-		err = json.Unmarshal([]byte(bufStr), &pk)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		if pk.MD5 != sqlMD5 {
-			// It is possible that the JSON MD5 field does not match the SQL MD5 field
-			// This is harmless in itself since we throw away the JSON field,
-			// but it may be a symptom of problems elsewhere, so log it.
-			log.Warnf("inconsistent MD5 in database (sql=%s, json=%s), ignoring json", sqlMD5, pk.MD5)
-		}
-
-		rfp := openpgp.Reverse(pk.Fingerprint)
-		key, err := readOneKey(pk.Bytes(), rfp)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		if autoPreen {
-			err = st.preen(key, pk, sqlMD5)
-			if err != nil {
-				continue
-			}
-		}
-		result = append(result, key)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return result, nil
-}
-
-func (st *storage) FetchRecords(rfps []string, options ...string) ([]*hkpstorage.Record, error) {
-	autoPreen := slices.Contains(options, hkpstorage.AutoPreen)
-	var rfpIn []string
-	for _, rfp := range rfps {
-		_, err := hex.DecodeString(rfp)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid rfingerprint %q", rfp)
-		}
-		rfpIn = append(rfpIn, "'"+strings.ToLower(rfp)+"'")
-	}
-	sqlStr := fmt.Sprintf("SELECT doc, md5, ctime, mtime FROM keys WHERE rfingerprint IN (%s)", strings.Join(rfpIn, ","))
-	rows, err := st.Query(sqlStr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	var result []*hkpstorage.Record
-	defer rows.Close()
-	for rows.Next() {
-		var bufStr, sqlMD5 string
-		var record hkpstorage.Record
-		err = rows.Scan(&bufStr, &sqlMD5, &record.CTime, &record.MTime)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		var pk jsonhkp.PrimaryKey
-		err = json.Unmarshal([]byte(bufStr), &pk)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		if pk.MD5 != sqlMD5 {
-			// It is possible that the JSON MD5 field does not match the SQL MD5 field
-			// This is harmless in itself since we throw away the JSON field,
-			// but it may be a symptom of problems elsewhere, so log it.
-			log.Warnf("inconsistent MD5 in database (sql=%s, json=%s), ignoring json", sqlMD5, pk.MD5)
-		}
-
-		rfp := openpgp.Reverse(pk.Fingerprint)
-		key, err := readOneKey(pk.Bytes(), rfp)
-		if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		if autoPreen {
-			err = st.preen(key, pk, sqlMD5)
-			if err != nil {
-				continue
-			}
-		}
-		record.PrimaryKey = key
-		result = append(result, &record)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return result, nil
-}
-
-func (st *storage) fetchKeydocs(rfps []string) ([]*keyDoc, error) {
-	var rfpIn []string
-	for _, rfp := range rfps {
-		_, err := hex.DecodeString(rfp)
-		if err != nil {
-			return nil, errors.Wrapf(err, "invalid rfingerprint %q", rfp)
-		}
-		rfpIn = append(rfpIn, "'"+strings.ToLower(rfp)+"'")
-	}
-	sqlStr := fmt.Sprintf("SELECT rfingerprint, doc, md5, ctime, mtime, idxtime, keywords FROM keys WHERE rfingerprint IN (%s)", strings.Join(rfpIn, ","))
-	rows, err := st.Query(sqlStr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	var result []*keyDoc
-	defer rows.Close()
-	for rows.Next() {
-		var kd keyDoc
-		err = rows.Scan(&kd.RFingerprint, &kd.Doc, &kd.MD5, &kd.CTime, &kd.MTime, &kd.IdxTime, &kd.Keywords)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		result = append(result, &kd)
-	}
-	err = rows.Err()
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	return result, nil
 }
 
 func readOneKey(b []byte, rfingerprint string) (*openpgp.PrimaryKey, error) {
@@ -785,633 +205,6 @@ func readOneKey(b []byte, rfingerprint string) (*openpgp.PrimaryKey, error) {
 			rfingerprint, keys[0].RFingerprint)
 	}
 	return keys[0], nil
-}
-
-func (st *storage) upsertKeyOnInsert(pubkey *openpgp.PrimaryKey) (kc hkpstorage.KeyChange, err error) {
-	var lastKey *openpgp.PrimaryKey
-	// Use AutoPreen even though it may cause double-update, because FetchKeys discards sqlMD5 so we can't examine it.
-	lastKeys, err := st.FetchKeys([]string{pubkey.RFingerprint}, hkpstorage.AutoPreen)
-	if err == nil {
-		// match primary fingerprint -- someone might have reused a subkey somewhere
-		err = hkpstorage.ErrKeyNotFound
-		for _, key := range lastKeys {
-			if key.RFingerprint == pubkey.RFingerprint {
-				lastKey, err = key, nil
-				break
-			}
-		}
-	}
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	if pubkey.UUID != lastKey.UUID {
-		return nil, errors.Errorf("upsert key %q lookup failed, found mismatch %q", pubkey.UUID, lastKey.UUID)
-	}
-	lastID := lastKey.KeyID()
-	lastMD5 := lastKey.MD5
-	err = openpgp.Merge(lastKey, pubkey)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	if lastMD5 != lastKey.MD5 {
-		err = st.Update(lastKey, lastID, lastMD5)
-		if err == errTargetMissing {
-			// propagate verbatim so it can be handled
-			return nil, err
-		} else if err != nil {
-			return nil, errors.WithStack(err)
-		}
-		return hkpstorage.KeyReplaced{OldID: lastID, OldDigest: lastMD5, NewID: lastKey.KeyID(), NewDigest: lastKey.MD5}, nil
-	}
-	return hkpstorage.KeyNotChanged{ID: lastID, Digest: lastMD5}, nil
-}
-
-func (st *storage) insertKey(key *openpgp.PrimaryKey) (needUpsert bool, retErr error) {
-	tx, err := st.Begin()
-	if err != nil {
-		return false, errors.WithStack(err)
-	}
-	defer func() {
-		if retErr != nil {
-			tx.Rollback()
-		} else {
-			tx.Commit()
-		}
-	}()
-	return st.insertKeyTx(tx, key)
-}
-
-func (st *storage) insertKeyTx(tx *sql.Tx, key *openpgp.PrimaryKey) (needUpsert bool, retErr error) {
-	stmt, err := tx.Prepare("INSERT INTO keys (rfingerprint, ctime, mtime, idxtime, md5, doc, keywords) " +
-		"SELECT $1::TEXT, $2::TIMESTAMP, $3::TIMESTAMP, $4::TIMESTAMP, $5::TEXT, $6::JSONB, $7::TSVECTOR " +
-		"WHERE NOT EXISTS (SELECT 1 FROM keys WHERE rfingerprint = $1)")
-	if err != nil {
-		return false, errors.WithStack(err)
-	}
-	defer stmt.Close()
-
-	subStmt, err := tx.Prepare("INSERT INTO subkeys (rfingerprint, rsubfp) " +
-		"SELECT $1::TEXT, $2::TEXT WHERE NOT EXISTS (SELECT 1 FROM subkeys WHERE rsubfp = $2)")
-	if err != nil {
-		return false, errors.WithStack(err)
-	}
-	defer subStmt.Close()
-
-	openpgp.Sort(key)
-
-	now := time.Now().UTC()
-	jsonKey := jsonhkp.NewPrimaryKey(key)
-	jsonBuf, err := json.Marshal(jsonKey)
-	if err != nil {
-		return false, errors.Wrapf(err, "cannot serialize rfp=%q", key.RFingerprint)
-	}
-
-	jsonStr := string(jsonBuf)
-	keywords := keywordsTSVector(key)
-	result, err := stmt.Exec(&key.RFingerprint, &now, &now, &now, &key.MD5, &jsonStr, &keywords)
-	if err != nil {
-		return false, errors.Wrapf(err, "cannot insert rfp=%q", key.RFingerprint)
-	}
-
-	var keysInserted int64
-	if keysInserted, err = result.RowsAffected(); err != nil {
-		// We arrive here if the DB driver doesn't support
-		// RowsAffected, although lib/pq is known to support it.
-		// If it doesn't, then something has gone badly awry!
-		return false, errors.Wrapf(err, "rows affected not available when inserting rfp=%q", key.RFingerprint)
-	}
-	if keysInserted == 0 {
-		return true, nil
-	}
-
-	for _, subKey := range key.SubKeys {
-		_, err := subStmt.Exec(&key.RFingerprint, &subKey.RFingerprint)
-		if err != nil {
-			return false, errors.Wrapf(err, "cannot insert rsubfp=%q", subKey.RFingerprint)
-		}
-	}
-	return false, nil
-}
-
-func (st *storage) bulkInsertGetStats(result *hkpstorage.InsertError) (int, int, int, int) {
-	var maxDups, minDups, keysInserted, subkeysInserted int
-	// Get Duplicate stats
-	err := st.QueryRow(bulkInsNumMinDups).Scan(&minDups)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update duplicate stats: %v", err)
-		minDups = 0
-	}
-	// In-file duplicates may be duplicates even if we insert a subkey for a key's rfp
-	// FIXME: This might be costly and could be removed...
-	err = st.QueryRow(bulkInsNumPossibleDups).Scan(&maxDups)
-	maxDups += minDups
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update duplicate stats: %v", err)
-		maxDups = 0
-	}
-	// Get keys/subkeys inserted
-	err = st.QueryRow(bulkInsertedKeysNum).Scan(&keysInserted)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update keys inserted stats: %v", err)
-		keysInserted = 0
-	}
-	err = st.QueryRow(bulkInsertedSubkeysNum).Scan(&subkeysInserted)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update subkeys inserted stats: %v", err)
-		subkeysInserted = 0
-	}
-	return maxDups, minDups, keysInserted, subkeysInserted
-}
-
-func (st *storage) bulkExecSingleTx(bulkJobString, jobDesc []string) (err error) {
-	// In single transaction
-	tx, err := st.Begin()
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	defer func() {
-		if err == nil {
-			err = tx.Commit()
-		}
-		if err != nil {
-			tx.Rollback()
-		}
-	}()
-
-	for i := 0; i < len(bulkJobString); i++ {
-		bulkTxStmt, err := tx.Prepare(bulkJobString[i])
-		if err != nil {
-			return errors.Wrapf(err, "preparing DB server job %s", jobDesc[i])
-		}
-		defer bulkTxStmt.Close()
-		_, err = bulkTxStmt.Exec()
-		if err != nil {
-			return errors.Wrapf(err, "issuing DB server job %s", jobDesc[i])
-		}
-	}
-	return err
-}
-
-func (st *storage) bulkInsertCheckSubkeys(result *hkpstorage.InsertError) (nullTuples int, ok bool) {
-	// NULLs stats
-	var numNulls int
-	err := st.QueryRow(bulkInsNumNullSubkeys).Scan(&numNulls)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update subkeys with NULL stats: %v", err)
-	}
-
-	// (1) Itermediate insert: no NULL fields & no Duplicates (in-file or in DB)
-	// (2) Keep only subkeys with Duplicates in subkeys_copyin:
-	//     Delete 1st-stage checked subkeys above & those with NULL fields
-	// (3) Single-copy of in-file Dups but not in-DB Dups
-	txStrs := []string{bulkTxFilterUniqueSubkeys, bulkTxPrepSubkeyStats, bulkTxFilterDupSubkeys}
-	msgStrs := []string{"bulkTx-filter-unique-subkeys", "bulkTx-prep-subkeys-stats", "bulkTx-filter-dup-subkeys"}
-	err = st.bulkExecSingleTx(txStrs, msgStrs)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return 0, false
-	}
-	return numNulls, true
-}
-
-func (st *storage) bulkInsertCheckKeys(result *hkpstorage.InsertError) (n int, ok bool) {
-	// NULLs stats
-	var numNulls int
-	err := st.QueryRow(bulkInsNumNullKeys).Scan(&numNulls)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update keys with NULL stats: %v", err)
-	}
-
-	// (1) rfingerprint & md5 are also UNIQUE in keys_checked so no duplicates inside this same file allowed
-	// (2) Keep only keys with Duplicates in keys_copyin: delete 1st-stage checked keys & tuples with NULL fields
-	// (3) Insert single copy of in-file Duplicates, if they have no Duplicate in final keys table (in DB)
-	txStrs := []string{bulkTxFilterUniqueKeys, bulkTxPrepKeyStats, bulkTxFilterDupKeys}
-	msgStrs := []string{"bulkTx-filter-unique-keys", "bulkTx-prep-key-stats", "bulkTx-filter-dup-keys"}
-	err = st.bulkExecSingleTx(txStrs, msgStrs)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return 0, false
-	}
-	return numNulls, true
-}
-
-func (st *storage) bulkInsertCheckedKeysSubkeys(result *hkpstorage.InsertError) (nullKeys, nullSubkeys int, ok bool) {
-	keysOK, subkeysOK := true, true
-	// key batch-processing
-	if nullKeys, keysOK = st.bulkInsertCheckKeys(result); !keysOK {
-		return 0, 0, false
-	}
-	// subkey batch-processing
-	if nullSubkeys, subkeysOK = st.bulkInsertCheckSubkeys(result); !subkeysOK {
-		return 0, 0, false
-	}
-
-	// Batch INSERT all checked-for-constraints keys from memory tables (should need no checks!!!!)
-	// Final batch-insertion in keys/subkeys tables without any checks: _must not_ give any errors
-	txStrs := []string{bulkTxInsertKeys, bulkTxInsertSubkeys}
-	msgStrs := []string{"bulkTx-insert-keys", "bulkTx-insert-subkeys"}
-	err := st.bulkExecSingleTx(txStrs, msgStrs)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return 0, 0, false
-	}
-	return nullKeys, nullSubkeys, true
-}
-
-func (st *storage) bulkInsertSendBunchTx(keystmt, msgSpec string, keysValueArgs []interface{}) (err error) {
-	// In single transaction...
-	tx, err := st.Begin()
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	defer func() {
-		if err != nil {
-			tx.Rollback()
-		} else {
-			tx.Commit()
-		}
-	}()
-
-	stmt, err := tx.Prepare(keystmt)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	defer stmt.Close()
-	_, err = stmt.Exec(keysValueArgs...) // All keys in bunch
-	if err != nil {
-		return errors.Wrapf(err, "cannot simply send a bunch of %s to server (too large bunch?)", msgSpec)
-	}
-	return nil
-}
-
-type keyInsertArgs struct {
-	RFingerprint *string
-	jsonStrDoc   *string
-	MD5          *string
-	keywords     *string
-}
-type subkeyInsertArgs struct {
-	keyRFingerprint    *string
-	subkeyRFingerprint *string
-}
-
-// Insert keys & subkeys to in-mem tables with no constraints at all: should have no errors!
-func (st *storage) bulkInsertDoCopy(keyInsArgs []keyInsertArgs, skeyInsArgs [][]subkeyInsertArgs,
-	result *hkpstorage.InsertError) (ok bool) {
-	lenKIA := len(keyInsArgs)
-	for idx, lastIdx := 0, 0; idx < lenKIA; lastIdx = idx {
-		totKeyArgs, totSubkeyArgs := 0, 0
-		keysValueStrings := make([]string, 0, keysInBunch)
-		keysValueArgs := make([]interface{}, 0, keysInBunch*keysNumColumns) // *** must be less than 64k arguments ***
-		subkeysValueStrings := make([]string, 0, subkeysInBunch)
-		subkeysValueArgs := make([]interface{}, 0, subkeysInBunch*subkeysNumColumns) // *** must be less than 64k arguments ***
-		for i, j := 0, 0; idx < lenKIA; idx, i = idx+1, i+1 {
-			lenSKIA := len(skeyInsArgs[idx])
-			totKeyArgs += keysNumColumns
-			totSubkeyArgs += subkeysNumColumns * lenSKIA
-			if (totKeyArgs > keysInBunch*keysNumColumns) || (totSubkeyArgs > subkeysInBunch*subkeysNumColumns) {
-				totKeyArgs -= keysNumColumns
-				totSubkeyArgs -= subkeysNumColumns * lenSKIA
-				break
-			}
-			keysValueStrings = append(keysValueStrings,
-				fmt.Sprintf("($%d::TEXT, $%d::JSONB, $%d::TIMESTAMP, $%d::TIMESTAMP, $%d::TIMESTAMP, $%d::TEXT, $%d::TSVECTOR)",
-					i*keysNumColumns+1, i*keysNumColumns+2, i*keysNumColumns+3, i*keysNumColumns+4, i*keysNumColumns+5, i*keysNumColumns+6, i*keysNumColumns+7))
-			insTime := time.Now().UTC()
-			keysValueArgs = append(keysValueArgs, *keyInsArgs[idx].RFingerprint, *keyInsArgs[idx].jsonStrDoc,
-				insTime, insTime, insTime, *keyInsArgs[idx].MD5, *keyInsArgs[idx].keywords)
-
-			for sidx := 0; sidx < lenSKIA; sidx, j = sidx+1, j+1 {
-				subkeysValueStrings = append(subkeysValueStrings, fmt.Sprintf("($%d::TEXT, $%d::TEXT)", j*subkeysNumColumns+1, j*subkeysNumColumns+2))
-				subkeysValueArgs = append(subkeysValueArgs,
-					*skeyInsArgs[idx][sidx].keyRFingerprint, *skeyInsArgs[idx][sidx].subkeyRFingerprint)
-			}
-		}
-		log.Debugf("attempting bulk insertion of %d keys and a total of %d subkeys!", idx-lastIdx, totSubkeyArgs>>1)
-
-		// Send all keys to in-mem tables to the pg server; *no constraints checked*
-		keystmt := fmt.Sprintf("INSERT INTO %s (rfingerprint, doc, ctime, mtime, idxtime, md5, keywords) VALUES %s",
-			keys_copyin_temp_table_name, strings.Join(keysValueStrings, ","))
-		err := st.bulkInsertSendBunchTx(keystmt, "keys", keysValueArgs)
-		if err != nil {
-			result.Errors = append(result.Errors, err)
-			return false
-		}
-
-		// Send all subkeys to in-mem tables to the pg server; *no constraints checked*
-		subkeystmt := fmt.Sprintf("INSERT INTO %s (rfingerprint, rsubfp) VALUES %s",
-			subkeys_copyin_temp_table_name, strings.Join(subkeysValueStrings, ","))
-		err = st.bulkInsertSendBunchTx(subkeystmt, "subkeys", subkeysValueArgs)
-		if err != nil {
-			result.Errors = append(result.Errors, err)
-			return false
-		}
-
-		log.Debugf("%d keys, %d subkeys sent to DB...", idx-lastIdx, totSubkeyArgs>>1)
-	}
-	return true
-}
-
-func (st *storage) bulkInsertCopyKeysToServer(keys []*openpgp.PrimaryKey, result *hkpstorage.InsertError) (int, bool) {
-	var key *openpgp.PrimaryKey
-	keyInsArgs := make([]keyInsertArgs, 0, len(keys))
-	skeyInsArgs := make([][]subkeyInsertArgs, 0, len(keys))
-	jsonStrs, theKeywords := make([]string, len(keys)), make([]string, len(keys))
-
-	unprocessed, sidx, i := 0, 0, 0
-	for _, key = range keys {
-		openpgp.Sort(key)
-		jsonKey := jsonhkp.NewPrimaryKey(key)
-		jsonBuf, err := json.Marshal(jsonKey)
-		if err != nil {
-			result.Errors = append(result.Errors,
-				errors.Wrapf(err, "pre-processing cannot serialize rfp=%q", key.RFingerprint))
-			unprocessed++
-			continue
-		}
-		jsonStrs[i], theKeywords[i] = string(jsonBuf), keywordsTSVector(key)
-		keyInsArgs = keyInsArgs[:i+1] // re-slice +1
-		keyInsArgs[i] = keyInsertArgs{&key.RFingerprint, &jsonStrs[i], &key.MD5, &theKeywords[i]}
-
-		skeyInsArgs = skeyInsArgs[:i+1] // re-slice +1
-		skeyInsArgs[i] = make([]subkeyInsertArgs, 0, len(key.SubKeys))
-		for sidx = 0; sidx < len(key.SubKeys); sidx++ {
-			skeyInsArgs[i] = skeyInsArgs[i][:sidx+1] // re-slice +1
-			skeyInsArgs[i][sidx] = subkeyInsertArgs{&key.RFingerprint, &key.SubKeys[sidx].RFingerprint}
-		}
-		i++
-	}
-	ok := st.bulkInsertDoCopy(keyInsArgs, skeyInsArgs, result)
-	return unprocessed, ok
-}
-
-func (st *storage) bulkDropTempTables() error {
-	// Drop the 2 pairs (all) of temporary tables
-	err := st.bulkExecSingleTx(drTempTablesSQL, sqlDesc(drTempTablesSQL))
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	return nil
-}
-
-func (st *storage) bulkCreateTempTables() error {
-	err := st.bulkExecSingleTx(crTempTablesSQL, sqlDesc(crTempTablesSQL))
-	if err != nil {
-		return errors.Wrap(err, "cannot drop temporary tables")
-	}
-	return nil
-}
-
-func (st *storage) BulkInsert(keys []*openpgp.PrimaryKey, result *hkpstorage.InsertError) (int, bool) {
-	log.Infof("attempting bulk insertion of keys")
-	t := time.Now() // FIXME: Remove this
-	// Create 2 pairs of _temporary_ (in-mem) tables:
-	// (a) keys_copyin, subkeys_copyin
-	// (b) keys_checked, subkeys_checked
-	err := st.bulkCreateTempTables()
-	if err != nil {
-		// This should always be possible (maybe, out-of-memory?)
-		result.Errors = append(result.Errors, err)
-		return 0, false
-	}
-	defer st.bulkDropTempTables()
-	keysWithNulls, subkeysWithNulls, ok := 0, 0, true
-	maxDups, minDups, keysInserted, subkeysInserted := 0, 0, 0, 0
-	// (a): Send *all* keys to in-mem tables on the pg server; *no constraints checked*
-	if _, ok = st.bulkInsertCopyKeysToServer(keys, result); !ok {
-		return 0, false
-	}
-	// (b): From _copyin tables (still only to in-mem table) remove duplicates
-	//      check *all* constraints & RollBack insertions of key/subkeys that err
-	if keysWithNulls, subkeysWithNulls, ok = st.bulkInsertCheckedKeysSubkeys(result); !ok {
-		return 0, false
-	}
-
-	maxDups, minDups, keysInserted, subkeysInserted = st.bulkInsertGetStats(result)
-	err = st.BulkNotify(bulkInsQueryKeyChange)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-	}
-
-	if minDups == maxDups {
-		log.Infof("%d keys and %d subkeys bulk-inserted, %d duplicates skipped (%d keys and %d subkeys with NULLs) in %v",
-			keysInserted, subkeysInserted, minDups, keysWithNulls, subkeysWithNulls, time.Since(t))
-	} else {
-		log.Infof("%d keys and %d subkeys bulk-inserted, at least %d (and up to %d possible) duplicates skipped "+
-			"(%d keys and %d subkeys with NULLs) in %v",
-			keysInserted, subkeysInserted, minDups, maxDups, keysWithNulls, subkeysWithNulls, time.Since(t))
-	}
-
-	err = st.bulkDropTempTables()
-	if err != nil {
-		// Temporary tables with previous data may lead to errors,
-		// when attempting insertion of duplicates, in next file,
-		// but may be resolved for the subsequent file(s)
-		result.Errors = append(result.Errors, err)
-	}
-	// FIXME: Imitate returning duplicates for reporting. Can be removed.
-	result.Duplicates = make([]*openpgp.PrimaryKey, minDups)
-	return keysInserted, true
-}
-
-var errTargetMissing = errors.New("errTargetMissing")
-
-func (st *storage) Insert(keys []*openpgp.PrimaryKey) (u, n int, retErr error) {
-	var result hkpstorage.InsertError
-
-	bulkOK, bulkSkip := false, false
-	if len(keys) >= minKeys2UseBulk {
-		// Attempt bulk insertion
-		n, bulkOK = st.BulkInsert(keys, &result)
-	} else {
-		bulkSkip = true
-	}
-
-	if !bulkOK {
-		log.Infof("bulk insertion %s; reverting to normal insertion",
-			(map[bool]string{true: "skipped (small number of keys)", false: "failed"})[bulkSkip])
-
-		for _, key := range keys {
-			if count, max := len(result.Errors), maxInsertErrors; count > max {
-				result.Errors = append(result.Errors,
-					errors.Errorf("too many insert errors (%d > %d), bailing...", count, max))
-				return u, n, result
-			}
-
-			if needUpsert, err := st.insertKey(key); err != nil {
-				result.Errors = append(result.Errors, err)
-				continue
-			} else if needUpsert {
-				var kc hkpstorage.KeyChange
-				// errTargetMissing is thrown if Update() can't find the key it was told to modify.
-				// This can happen in case of concurrent updates to the same key. Back off a few times.
-				for i := 0; i < 3; i++ {
-					kc, err = st.upsertKeyOnInsert(key)
-					if err != errTargetMissing {
-						log.Infof("key fp(%v) is slippery; backing off", key.Fingerprint())
-						break
-					}
-				}
-				if err == errTargetMissing {
-					result.Errors = append(result.Errors,
-						errors.Errorf("key fp(%v) was changing while we were updating it", key.Fingerprint()))
-				} else if err != nil {
-					result.Errors = append(result.Errors, err)
-					continue
-				} else {
-					switch kc.(type) {
-					case hkpstorage.KeyReplaced:
-						// FIXME: Listener in hockeypuck-load not really prepared for
-						// hkpstorage.KeyReplaced notifications but stats are updated...
-						st.Notify(kc)
-						u++
-					case hkpstorage.KeyNotChanged:
-						result.Duplicates = append(result.Duplicates, key)
-					}
-				}
-				continue
-			} else {
-				st.Notify(hkpstorage.KeyAdded{
-					ID:     key.KeyID(),
-					Digest: key.MD5,
-				})
-				n++
-			}
-		}
-	}
-
-	if len(result.Duplicates) > 0 || len(result.Errors) > 0 {
-		return u, n, result
-	}
-	return u, n, nil
-}
-
-func (st *storage) Replace(key *openpgp.PrimaryKey) (_ string, retErr error) {
-	tx, err := st.Begin()
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-	defer func() {
-		if retErr != nil {
-			tx.Rollback()
-		} else {
-			retErr = tx.Commit()
-		}
-	}()
-	md5, err := st.deleteTx(tx, key.Fingerprint())
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-	_, err = st.insertKeyTx(tx, key)
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-
-	st.Notify(hkpstorage.KeyReplaced{
-		OldID:     key.KeyID(),
-		OldDigest: md5,
-		NewID:     key.KeyID(),
-		NewDigest: key.MD5,
-	})
-	return md5, nil
-}
-
-func (st *storage) Delete(fp string) (_ string, retErr error) {
-	tx, err := st.Begin()
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-	defer func() {
-		if retErr != nil {
-			tx.Rollback()
-		} else {
-			retErr = tx.Commit()
-		}
-	}()
-	md5, err := st.deleteTx(tx, fp)
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-	st.Notify(hkpstorage.KeyRemoved{ID: fp, Digest: md5})
-	return md5, nil
-}
-
-// deleteTx does not handle cleanup; the caller MUST defer commit/rollback
-func (st *storage) deleteTx(tx *sql.Tx, fp string) (string, error) {
-	rfp := openpgp.Reverse(fp)
-	_, err := tx.Exec("DELETE FROM subkeys WHERE rfingerprint = $1", rfp)
-	if err != nil {
-		return "", errors.WithStack(err)
-	}
-	var md5 string
-	err = tx.QueryRow("DELETE FROM keys WHERE rfingerprint = $1 RETURNING md5", rfp).Scan(&md5)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return "", errors.WithStack(hkpstorage.ErrKeyNotFound)
-		}
-		return "", errors.WithStack(err)
-	}
-	return md5, nil
-}
-
-func (st *storage) Update(key *openpgp.PrimaryKey, lastID string, lastMD5 string) (retErr error) {
-	tx, err := st.Begin()
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	defer func() {
-		if retErr != nil {
-			tx.Rollback()
-		} else {
-			tx.Commit()
-		}
-	}()
-
-	openpgp.Sort(key)
-
-	now := time.Now().UTC()
-	jsonKey := jsonhkp.NewPrimaryKey(key)
-	jsonBuf, err := json.Marshal(jsonKey)
-	if err != nil {
-		return errors.Wrapf(err, "cannot serialize rfp=%q", key.RFingerprint)
-	}
-	keywords := keywordsTSVector(key)
-	result, err := tx.Exec("UPDATE keys SET mtime = $1, idxtime = $2, md5 = $3, keywords = $4::TSVECTOR, doc = $5 "+
-		"WHERE md5 = $6",
-		&now, &now, &key.MD5, &keywords, jsonBuf, lastMD5)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	rowsAffected, err := result.RowsAffected()
-	if err != nil || rowsAffected > 1 {
-		return errors.Errorf("unexpected error when updating digest %v fp(%v)", lastMD5, key.Fingerprint())
-	} else if rowsAffected == 0 {
-		// The md5 disappeared before we could update it. Thread-safety backoff.
-		return errTargetMissing
-	}
-	for _, subKey := range key.SubKeys {
-		_, err := tx.Exec("INSERT INTO subkeys (rfingerprint, rsubfp) "+
-			"SELECT $1::TEXT, $2::TEXT WHERE NOT EXISTS (SELECT 1 FROM subkeys WHERE rsubfp = $2)",
-			&key.RFingerprint, &subKey.RFingerprint)
-		if err != nil {
-			return errors.WithStack(err)
-		}
-	}
-
-	st.Notify(hkpstorage.KeyReplaced{
-		OldID:     lastID,
-		OldDigest: lastMD5,
-		NewID:     key.KeyID(),
-		NewDigest: key.MD5,
-	})
-	return nil
 }
 
 func keywordsTSVector(key *openpgp.PrimaryKey) string {
@@ -1626,367 +419,4 @@ func subkeys(key *openpgp.PrimaryKey) []string {
 		result = append(result, subkey.RFingerprint)
 	}
 	return result
-}
-
-func (st *storage) Subscribe(f func(hkpstorage.KeyChange) error) {
-	st.mu.Lock()
-	st.listeners = append(st.listeners, f)
-	st.mu.Unlock()
-}
-
-func (st *storage) Notify(change hkpstorage.KeyChange) error {
-	st.mu.Lock()
-	defer st.mu.Unlock()
-	log.Debugf("%v", change)
-	for _, f := range st.listeners {
-		err := f(change)
-		if err != nil {
-			log.Errorf("notify failed: %v", err)
-		}
-	}
-	return nil
-}
-
-func (st *storage) BulkNotify(sqlStr string) error {
-	rows, err := st.Query(sqlStr)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-
-	defer rows.Close()
-	for rows.Next() {
-		var md5 string
-		err := rows.Scan(&md5)
-		if err != nil {
-			if err == sql.ErrNoRows {
-				return nil
-			} else {
-				return errors.WithStack(err)
-			}
-		}
-		st.Notify(hkpstorage.KeyAdded{Digest: md5})
-	}
-	err = rows.Err()
-	return errors.WithStack(err)
-}
-
-func (st *storage) RenotifyAll() error {
-	return st.BulkNotify("SELECT md5 FROM keys")
-}
-
-func (st *storage) bulkReindexDoCopy(keyDocs iter.Seq[*keyDoc], result *hkpstorage.InsertError) bool {
-	keyDocsPull, keyDocsPullStop := iter.Pull(keyDocs)
-	defer keyDocsPullStop()
-	pullOk := true
-	var kd *keyDoc
-	for idx, lastIdx := 0, 0; pullOk; lastIdx = idx {
-		totKeyArgs := 0
-		keysValueStrings := make([]string, 0, keysInBunch)
-		keysValueArgs := make([]interface{}, 0, keysInBunch*keysNumColumns)
-		kd, pullOk = keyDocsPull()
-		if !pullOk {
-			return true
-		}
-		for i := 0; pullOk; idx, i = idx+1, i+1 {
-			totKeyArgs += keysNumColumns
-			if totKeyArgs > keysInBunch*keysNumColumns {
-				totKeyArgs -= keysNumColumns
-				break
-			}
-			keysValueStrings = append(keysValueStrings,
-				fmt.Sprintf("($%d::TEXT, $%d::JSONB, $%d::TIMESTAMP, $%d::TIMESTAMP, $%d::TIMESTAMP, $%d::TEXT, $%d::TSVECTOR)",
-					i*keysNumColumns+1, i*keysNumColumns+2, i*keysNumColumns+3, i*keysNumColumns+4, i*keysNumColumns+5, i*keysNumColumns+6, i*keysNumColumns+7))
-			insTime := time.Now().UTC()
-			keysValueArgs = append(keysValueArgs, kd.RFingerprint, "{}",
-				insTime, insTime, insTime, kd.MD5, kd.Keywords)
-			kd, pullOk = keyDocsPull()
-		}
-		log.Debugf("attempting bulk copy of %d keys", idx-lastIdx)
-		keystmt := fmt.Sprintf("INSERT INTO %s (rfingerprint, doc, ctime, mtime, idxtime, md5, keywords) VALUES %s",
-			keys_copyin_temp_table_name, strings.Join(keysValueStrings, ","))
-
-		err := st.bulkInsertSendBunchTx(keystmt, "reindexes", keysValueArgs)
-		if err != nil {
-			result.Errors = append(result.Errors, err)
-			return false
-		}
-		log.Debugf("%d updates sent to DB...", idx-lastIdx)
-	}
-	return true
-}
-
-func (st *storage) bulkReindexGetStats(result *hkpstorage.InsertError) int {
-	var keysReindexed int
-	err := st.QueryRow(bulkCopiedKeysNum).Scan(&keysReindexed)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		log.Warnf("could not update reindex stats: %v", err)
-		keysReindexed = 0
-	}
-	return keysReindexed
-}
-
-func (st *storage) bulkReindexKeys(result *hkpstorage.InsertError) bool {
-	log.Debugf("attempting bulk update of keys")
-	txStrs := []string{bulkTxReindexKeys}
-	msgStrs := []string{"bulkTx-reindex-keys"}
-	err := st.bulkExecSingleTx(txStrs, msgStrs)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return false
-	}
-	return true
-}
-
-func (st *storage) bulkReindex(keyDocs map[string]*keyDoc, result *hkpstorage.InsertError) (int, bool) {
-	log.Infof("attempting bulk reindex of %d keys", len(keyDocs))
-	// We only use the `keys_copyin` temp table, but reuse the full complement for simplicity.
-	err := st.bulkCreateTempTables()
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return 0, false
-	}
-	defer st.bulkDropTempTables()
-	keysReindexed := 0
-	if !st.bulkReindexDoCopy(maps.Values(keyDocs), result) {
-		return 0, false
-	}
-	if !st.bulkReindexKeys(result) {
-		return 0, false
-	}
-
-	keysReindexed = st.bulkReindexGetStats(result)
-	err = st.bulkDropTempTables()
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-	}
-	return keysReindexed, true
-}
-
-// refresh updates the keyDoc fields that cache values from the jsonb document.
-// This is called by refreshBunch to ensure the DB columns are correctly populated,
-// for example after changes to the keyword indexing policy, or to the DB schema.
-func (kd *keyDoc) refresh() (changed bool, err error) {
-	// Unmarshal the doc
-	var pk jsonhkp.PrimaryKey
-	err = json.Unmarshal([]byte(kd.Doc), &pk)
-	if err != nil {
-		return false, err
-	}
-	rfp := openpgp.Reverse(pk.Fingerprint)
-	key, err := readOneKey(pk.Bytes(), rfp)
-	if err != nil {
-		return false, err
-	}
-	if key == nil {
-		// This should never happen, but has been observed in testing.
-		// Something is corrupt in the DB!
-		log.Errorf("nil returned successfully from readOneKey(%v)", pk)
-		return false, errors.Errorf("nil returned successfully from readOneKey(%v)", pk)
-	}
-
-	// Regenerate keywords
-	newKeywords, _, _ := keywordsFromKey(key)
-	oldKeywords := keywordsFromTSVector(kd.Keywords)
-	slices.Sort(newKeywords)
-	slices.Sort(oldKeywords)
-	if !slices.Equal(oldKeywords, newKeywords) {
-		log.Debugf("keyword mismatch on fp=%s, was %q now %q", pk.Fingerprint, oldKeywords, newKeywords)
-		kd.Keywords, err = keywordsToTSVector(newKeywords, " ")
-		changed = true
-	}
-
-	// In future we may add further tasks here.
-	// DO NOT update the md5 field, as this is used by bulkReindex to prevent simultaneous updates.
-
-	return changed, err
-}
-
-// refreshBunch fetches a bunch of keyDocs from the DB and returns freshened copies of the ones with stale records.
-// TODO: ModifiedSince habitually yields the same entries multiple times (FIXME!),
-// so we use a map (not an array) to deduplicate the returned keyDocs.
-func (st *storage) refreshBunch(bookmark *time.Time, newKeyDocs map[string]*keyDoc, result *hkpstorage.InsertError) (count int, finished bool) {
-	// ModifiedSince uses LIMIT, so this is safe
-	rfps, err := st.ModifiedSince(*bookmark)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return 0, false
-	}
-	if len(rfps) == 0 {
-		return 0, true
-	}
-	keyDocs, err := st.fetchKeydocs(rfps)
-	if err != nil {
-		result.Errors = append(result.Errors, err)
-		return 0, false
-	}
-	count = len(keyDocs)
-	log.Debugf("reindexing %d records", count)
-	for _, kd := range keyDocs {
-		*bookmark = kd.MTime
-		changed, err := kd.refresh()
-		if err != nil {
-			result.Errors = append(result.Errors, err)
-		} else if changed {
-			newKeyDocs[kd.MD5] = kd
-		}
-	}
-	log.Infof("found %d stale records up to %v", len(newKeyDocs), bookmark)
-	return count, false
-}
-
-// Reindex is a goroutine that reindexes the keydb in-place, oldest items first.
-// It does not update CTime, MTime, MD5 or Doc, and does not call Notify.
-// It always returns nil, as reindex failure is not fatal.
-func (st *storage) Reindex() error {
-	bookmark := time.Time{}
-	newKeyDocs := make(map[string]*keyDoc, keysInBunch)
-	result := hkpstorage.InsertError{}
-	total := 0
-
-	for {
-		select {
-		case <-st.t.Dying():
-			return nil
-		default:
-		}
-
-		t := time.Now()
-		count, finished := st.refreshBunch(&bookmark, newKeyDocs, &result)
-		total += count
-		if finished || len(newKeyDocs) > keysInBunch-100 {
-			n, bulkOK := st.bulkReindex(newKeyDocs, &result)
-			if !bulkOK {
-				log.Debugf("bulkReindex not ok, result: %q", result)
-				if count, max := len(result.Errors), maxInsertErrors; count > max {
-					log.Errorf("too many reindexing errors (%d > %d), bailing...", count, max)
-					return nil
-				}
-			}
-			log.Infof("%d keys reindexed in %v; total scanned %d", n, time.Since(t), total)
-			newKeyDocs = make(map[string]*keyDoc, keysInBunch)
-		}
-		if finished {
-			log.Infof("reindexing complete")
-			return nil
-		}
-	}
-}
-
-// Start reindexing in the background. This should only be done after server startup, not during load or dump.
-func (st *storage) StartReindex() {
-	st.t.Go(st.Reindex)
-}
-
-//
-// pks.Storage implementation
-//
-
-// Initialise a new PKS peer record if it does not already exist.
-func (st *storage) PKSInit(addr string, lastSync time.Time) error {
-	stmt, err := st.Prepare("INSERT INTO pks_status ( addr, last_sync, last_error ) VALUES ( $1, $2, $3 ) ON CONFLICT DO NOTHING")
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	_, err = stmt.Exec(addr, lastSync, sql.NullString{Valid: false})
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	return nil
-}
-
-// Return the status of all PKS peers.
-func (st *storage) PKSAll() ([]*pksstorage.Status, error) {
-	rows, err := st.Query("SELECT addr, last_sync, last_error FROM pks_status")
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	var result []*pksstorage.Status
-	defer rows.Close()
-	for rows.Next() {
-		var addr string
-		var lastErrorString sql.NullString
-		var lastError error
-		var lastSync time.Time
-		err = rows.Scan(&addr, &lastSync, &lastErrorString)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		if lastErrorString.Valid {
-			lastError = errors.Errorf(lastErrorString.String)
-		}
-		result = append(result, &pksstorage.Status{
-			Addr:      addr,
-			LastSync:  lastSync,
-			LastError: lastError,
-		})
-	}
-	return result, nil
-}
-
-// Get one PKS peer.
-func (st *storage) PKSGet(addr string) (*pksstorage.Status, error) {
-	stmt, err := st.Prepare("SELECT addr, last_sync, last_error FROM pks_status WHERE addr = $1")
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-	rows, err := stmt.Query(addr)
-	if err != nil {
-		return nil, errors.WithStack(err)
-	}
-
-	var result *pksstorage.Status
-	defer rows.Close()
-	for rows.Next() {
-		var addr string
-		var lastErrorString sql.NullString
-		var lastError error
-		var lastSync time.Time
-		err = rows.Scan(&addr, &lastSync, &lastErrorString)
-		if err != nil && err != sql.ErrNoRows {
-			return nil, errors.WithStack(err)
-		}
-		if lastErrorString.Valid {
-			lastError = errors.Errorf(lastErrorString.String)
-		}
-		result = &pksstorage.Status{
-			Addr:      addr,
-			LastSync:  lastSync,
-			LastError: lastError,
-		}
-		// Only process the first result; the storage SHOULD NOT contain duplicate records
-		break
-	}
-	return result, nil
-}
-
-// Update the status of one PKS peer.
-func (st *storage) PKSUpdate(status *pksstorage.Status) error {
-	stmt, err := st.Prepare("UPDATE pks_status SET last_sync = $2, last_error = $3 WHERE addr = $1")
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	lastErrorString := sql.NullString{Valid: false}
-	if status.LastError != nil {
-		lastErrorString = sql.NullString{String: status.LastError.Error(), Valid: true}
-	}
-	_, err = stmt.Exec(status.Addr, status.LastSync, lastErrorString)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	return nil
-}
-
-// Remove a PKS peer.
-func (st *storage) PKSRemove(addr string) error {
-	stmt, err := st.Prepare("DELETE FROM pks_status WHERE addr = $1")
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	_, err = stmt.Exec(&addr)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	return nil
 }
